backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Konrad Zagozda <zagozdakonrad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
django-allauth = "^0.54.0"
dj-rest-auth = "^3.0.0"
drf-spectacular = "^0.26.2"
celery = "^5.2.7"
django-celery-results = "^2.5.0"
django-celery-beat = "^2.5.0"
redis = "^4.5.4"

[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-django = "^4.5.2"
pytest-xdist = "^3.2.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings.test"
python_files = ["test_*.py", "*_test.py", "tests.py"]
addopts = "--cov=. --cov-report html"

[tool.coverage.report]
fail_under = 95

[tool.coverage.run]
omit = [
	"backend/settings/*",
	"backend/asgi.py",
	"backend/wsgi.py",
	"backend/celery.py",
	"**/urls.py",
	"manage.py",
	"**/tests.py",
	"**/migrations/**",
	"**/__init__.py",
	"**/tests/**",
	"conftest.py",
	"**/apps.py",
	"**/admin.py"
]
